chore: clippy and fmt done

+++ circuits/src/cpu/columns.rs
@@ -210,18 +210,26 @@ pub fn rangecheck_looking<F: Field>() -> Vec<Table<F>> {
         CpuTable::new(vec![cpu.abs_diff], &ops.bge + &ops.blt),
         CpuTable::new(vec![cpu.product_high_limb], muls.clone()),
         CpuTable::new(vec![cpu.product_low_limb], muls),
-        // range constraint `lv.op1_full_range() + lv.is_op1_signed() * CpuState::<P>::shifted(31);` as u32
-        CpuTable::new(vec![
-            &cpu.op1_value + 
-            &cpu.op1_sign_bit * F::from_canonical_u64((1 as u64) << 32).neg() + 
-            &is_op1_signed * F::from_canonical_u64((1 as u64) << 31)
-        ], is_running.clone()),
-        // range constraint `lv.op2_full_range() + lv.is_op2_signed() * CpuState::<P>::shifted(31);` as u32
-        CpuTable::new(vec![
-            &cpu.op2_value + 
-            &cpu.op2_sign_bit * F::from_canonical_u64((1 as u64) << 32).neg() +
-            &is_op2_signed * F::from_canonical_u64((1 as u64) << 31)
-        ], is_running),
+        // range constraint `lv.op1_full_range() + lv.is_op1_signed() *
+        // CpuState::<P>::shifted(31);` as u32
+        CpuTable::new(
+            vec![
+                &cpu.op1_value
+                    + &cpu.op1_sign_bit * F::from_canonical_u64((1_u64) << 32).neg()
+                    + &is_op1_signed * F::from_canonical_u64((1_u64) << 31),
+            ],
+            is_running.clone(),
+        ),
+        // range constraint `lv.op2_full_range() + lv.is_op2_signed() *
+        // CpuState::<P>::shifted(31);` as u32
+        CpuTable::new(
+            vec![
+                &cpu.op2_value
+                    + &cpu.op2_sign_bit * F::from_canonical_u64((1_u64) << 32).neg()
+                    + &is_op2_signed * F::from_canonical_u64((1_u64) << 31),
+            ],
+            is_running,
+        ),
     ]
 }
